news_project/requirements.txt
Django>=3.2
djangorestframework
mysqlclient
requests
